requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
rembg>=2.0.50
stripe>=8.0.0
requests>=2.31.0
numpy>=1.26.0
